+++ libraries/FreeIMU/examples/FreeIMU_yaw_pitch_roll/FreeIMU_yaw_pitch_roll.ino
@@ -46,7 +46,7 @@
   #define HAS_EEPPROM 1
 #endif
 
-int raw_values[11];
+int16_t raw_values[11];
 //char str[512];
 float ypr[3]; // yaw pitch roll
 float val[13];
